gpu-server/pyproject.toml
[project]
name = "lejonet-gpu"
version = "0.1.0"
description = "GPU inference server for Lejonet HTR"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "onnxruntime-gpu>=1.19",
    "numpy>=1.26",
    "pillow>=10",
    "huggingface-hub>=0.27",
    "ray[serve]>=2.40",
]

[project.optional-dependencies]
rocm = [
    "onnxruntime-rocm>=1.19",
]
dev = [
    "pytest>=8",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lejonet_gpu"]
